refactor: Migrate auth models to its own package
