+++ cmd/api-server/models/auth.go
@@ -1,4 +1,4 @@
-package auth
+package models
 
 import (
 	"context"
