+++ cmd/api-server/models/auth_test.go
@@ -1,4 +1,4 @@
-package auth
+package models
 
 import "testing"
